pyproject.toml
[project]
name = "aider-mcp"
version = "0.1.0"
description = "MCP server for integrating Aider AI coding tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aider-chat>=0.82.2",
    "fastmcp>=2.2.1",
    "pydantic>=2.11.3"
]

[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
aider-mcp-server = "aider_mcp_server.__main__:main"